feat(financials): implement multi-period projections with domain invariants and precision hardening
**Backend Improvements:**
- Refactor monetary arithmetic from double to BigDecimal across FinancialsService for
  floating-point precision and correctness in financial calculations
- Introduce ApiExceptionHandler for centralized, structured error handling with RFC 7807
  ProblemDetail responses
- Add FinancialsControllerTests validating API contract and error scenarios

**Frontend Enhancements:**
- Introduce financialsAnchors module defining domain invariants (rent withdrawal, rent reserve,
  primary paycheck) with type-safe predicates
- Implement three data integrity functions with migration support:
  - ensureRentWithdrawal: Auto-normalizes legacy rent entries or creates Rent withdrawal
  - ensureRentReserveAccount: Auto-creates rent reserve savings account in Cash & Savings category
  - ensurePrimaryPaycheck: Establishes "Net Income / Bi-Weekly" income summary entry
- Build financialsProjection module: multi-period cash flow calculation engine with rent coverage
  tracking and bill/withdrawal scheduling (156 lines + 245-line test suite)

**UI & Accessibility:**
- Refactor MonthlyWithdrawalsTab with improved table semantics and accessibility
- Update OverviewTab metric from "Bi-weekly disposable" to "Primary paycheck" with semantic table
  footer restructuring
- Add colgroup and semantic CSS classes for enhanced screen reader support across financials tables

**Quality & Documentation:**
- Add Vite coverage thresholds (statements 45%, branches 45%, functions 35%, lines 46%) to prevent
  test coverage regressions
- Document production readiness roadmap and architectural decisions in ADRs

+++ backend/src/main/java/com/example/backend/dto/financials/AssetAccountResponse.java
@@ -1,3 +1,5 @@
 package com.example.backend.dto.financials;
 
-public record AssetAccountResponse(long id, String account, String company, double amount) {}
+import java.math.BigDecimal;
+
+public record AssetAccountResponse(long id, String account, String company, BigDecimal amount) {}
