+++ backend/src/main/java/com/example/backend/dto/financials/AssetCategoryResponse.java
@@ -1,6 +1,7 @@
 package com.example.backend.dto.financials;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record AssetCategoryResponse(
-    String key, String label, double total, List<AssetAccountResponse> accounts) {}
+    String key, String label, BigDecimal total, List<AssetAccountResponse> accounts) {}
