+++ backend/src/main/java/com/example/backend/dto/financials/DebtAccountResponse.java
@@ -1,3 +1,5 @@
 package com.example.backend.dto.financials;
 
-public record DebtAccountResponse(long id, String account, String company, double amount) {}
+import java.math.BigDecimal;
+
+public record DebtAccountResponse(long id, String account, String company, BigDecimal amount) {}
