+++ backend/src/main/java/com/example/backend/repository/DebtAccount.java
@@ -1,6 +1,8 @@
 package com.example.backend.repository;
 
-public record DebtAccount(long id, String account, String company, double amount) {
+import java.math.BigDecimal;
+
+public record DebtAccount(long id, String account, String company, BigDecimal amount) {
 
   public DebtAccount withId(long id) {
     return new DebtAccount(id, account, company, amount);
